# solutions/pocito/marketplace-server/marketplace-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "marketplace",
    "version": "0.1.0",
    "description": "Requests are scoped by the x-wonder-room header; omitting it uses the marketplace room."
  },
  "paths": {
...
    "/api/v1/tools/{id}/code/{path}": {
      "get": {
        "tags": [
          "tools"
        ],
        "summary": "Tool Code",
        "description": "Read a tool source file. 404 if the object is missing.",
        "operationId": "get_tool_code_api_v1_tools__id__code__path__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/skills/{id}/SKILL.md": {
      "get": {
        "tags": [
          "skills"
        ],
        "summary": "Skill Md",
        "description": "Read the SKILL.md document. 404 if missing.",
        "operationId": "get_skill_md_api_v1_skills__id__SKILL_md_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/skills/{id}/assets/{path}": {
      "get": {
        "tags": [
          "skills"
        ],
        "summary": "Skill Asset",
        "description": "Read a skill asset as raw bytes (binary-safe). 404 if missing.",
        "operationId": "get_skill_asset_api_v1_skills__id__assets__path__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/plugins/{id}/config.yaml": {
      "get": {
        "tags": [
          "plugins"
        ],
        "summary": "Plugin Config",
        "description": "Read the config.yaml document. 404 if missing.",
        "operationId": "get_plugin_config_api_v1_plugins__id__config_yaml_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/plugins/{id}/README.md": {
      "get": {
        "tags": [
          "plugins"
        ],
        "summary": "Plugin Readme",
        "description": "Read the README.md document. 404 if missing.",
        "operationId": "get_plugin_readme_api_v1_plugins__id__README_md_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/plugins/{id}/references": {
      "get": {
        "tags": [
          "plugins"
        ],
        "summary": "Plugin References",
        "description": "Check existence of every tool/skill referenced by the plugin. 404 if missing.",
        "operationId": "check_plugin_references_api_v1_plugins__id__references_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Check Plugin References Api V1 Plugins  Name  References Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/agents/{id}/config.yaml": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "Agent Config",
        "description": "Read the config.yaml document. 404 if missing.",
        "operationId": "get_agent_config_api_v1_agents__id__config_yaml_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/{id}/references": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "Agent References",
        "description": "Check existence of every plugin/tool/skill/sub-agent. 404 if missing.",
        "operationId": "check_agent_references_api_v1_agents__id__references_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Check Agent References Api V1 Agents  Name  References Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/presign/download": {
      "post": {
        "tags": [
          "presign"
        ],
        "summary": "Presign Download",
        "description": "Return a presigned GET URL for payload.key.",
        "operationId": "presign_download_api_v1_presign_download_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Presign Download Api V1 Presign Download Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/presign/upload": {
      "post": {
        "tags": [
          "presign"
        ],
        "summary": "Presign Upload",
        "description": "Return a presigned PUT URL for payload.key.",
        "operationId": "presign_upload_api_v1_presign_upload_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Presign Upload Api V1 Presign Upload Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/audit/{resource_type}/{resource_id}": {
      "get": {
        "tags": [
          "audit"
        ],
        "summary": "Audit",
        "description": "List audit events for resource_id/resource_type; corrupt events are skipped and missing resources return an empty list.",
        "operationId": "list_audit_events_api_v1_audit__resource_type___resource_id__get",
        "parameters": [
          {
            "name": "resource_type",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "tool",
                "skill",
                "plugin",
                "agent",
                "knowledge"
              ],
              "type": "string",
              "title": "Resource Type"
            }
          },
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resource Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Audit Events Api V1 Audit  Resource Type   Resource Name  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Create User",
        "description": "Create a user. Returns the new UserObject. uids are unique; usernames are not.",
        "operationId": "create_user_api_v1_users__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUser"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{uid}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get User",
        "description": "Read a user by id. 404 if missing; 422 if the stored object is corrupt.",
        "operationId": "get_user_api_v1_users__uid__get",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Uid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get User Api V1 Users  Uid  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DownloadRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "key"
        ],
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "expires_in": {
            "type": "integer",
            "minimum": 1.0,
            "maximum": 86400.0,
            "default": 3600,
            "title": "Expires In"
          }
        },
        "title": "DownloadRequest",
        "description": "Request body for a presigned download URL."
      },
      "HTTPValidationError": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail"
          }
        },
        "title": "HTTPValidationError"
      },
...
      "UploadRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "key"
        ],
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "content_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Type"
          },
          "expires_in": {
            "type": "integer",
            "minimum": 1.0,
            "maximum": 86400.0,
            "default": 3600,
            "title": "Expires In"
          }
        },
        "title": "UploadRequest",
        "description": "Request body for a presigned upload URL."
      },
...
    }
  }
}